Finalize e-bike charger firmware with comprehensive verification and architectural improvements.
- Modularized firmware into object-oriented classes (Charger, UI, Storage, Filter).
- Implemented robust incremental CC/CV charging strategy.
- Integrated ATX PSU control with 20s anti-cycle cooldown.
- Added comprehensive safety suite: over-temp, over-current, timeout, and disconnect detection.
- Developed 6-screen OLED interface with live telemetry and voltage graph.
- Implemented high-precision 'double' integration for Ah/Wh logging.
- Enhanced ADC accuracy using millivolt-based readings (analogReadMilliVolts).
- Provided 5-suite automated test framework in C++ mock environment.
- Optimized Flash storage persistence with significance-based wear leveling.
- Documented system architecture and testing in README.md.

+++ Storage.h
@@ -18,24 +18,24 @@ public:
 #endif
     }
 
-    void save(float ah, float wh) {
+    void save(double ah, double wh) {
         _lifetimeAh = ah;
         _lifetimeWh = wh;
 #if defined(ARDUINO_ARCH_ESP32) || defined(ESP32)
-        _prefs.putFloat("ah", _lifetimeAh);
-        _prefs.putFloat("wh", _lifetimeWh);
+        _prefs.putFloat("ah", (float)_lifetimeAh);
+        _prefs.putFloat("wh", (float)_lifetimeWh);
 #endif
     }
 
-    float lifetimeAh() const { return _lifetimeAh; }
-    float lifetimeWh() const { return _lifetimeWh; }
+    double lifetimeAh() const { return _lifetimeAh; }
+    double lifetimeWh() const { return _lifetimeWh; }
 
 private:
 #if defined(ARDUINO_ARCH_ESP32) || defined(ESP32)
     Preferences _prefs;
 #endif
-    float _lifetimeAh;
-    float _lifetimeWh;
+    double _lifetimeAh;
+    double _lifetimeWh;
 };
 
 #endif
